Alfa

+++ application/controllers/Suppliers.php
@@ -117,6 +117,13 @@ class Suppliers extends CI_Controller {
 	public function delete()
 	{
 		$id = $this->input->post('supplier_id');
+		$material = $this->Supplier_model->get_material($id)->row();
+
+		if( $material != null ) {
+			$this->session->set_flashdata('used', 'Data sedang digunakan di daftar bahan, silahkan hapus semua data bahan yang berhubungan dengan kategori ini.');
+			redirect('suppliers');
+		}
+
 		$this->Supplier_model->delete($id);
 
 		$this->session->set_flashdata('deleted', 'Data berhasil dihapus.');
